+++ benchmarks/battleship/app/src/main/java/com/android/designcompose/benchmarks/battleship/MainActivity.kt
@@ -20,6 +20,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import com.android.designcompose.DesignSettings
+import com.android.designcompose.benchmarks.battleship.lib.BattleshipTest
 import com.android.designcompose.test.internal.interFont
 
 class MainActivity : ComponentActivity() {
