+++ integration-tests/validation/src/main/java/com/android/designcompose/testapp/validation/examples/AllExamples.kt
@@ -17,6 +17,8 @@
 package com.android.designcompose.testapp.validation.examples
 
 import androidx.compose.runtime.Composable
+import com.android.designcompose.benchmarks.battleship.lib.BattleshipDoc
+import com.android.designcompose.benchmarks.battleship.lib.BattleshipTest
 
 val EXAMPLES: ArrayList<Triple<String, @Composable () -> Unit, String?>> =
     arrayListOf(
@@ -26,7 +28,7 @@ val EXAMPLES: ArrayList<Triple<String, @Composable () -> Unit, String?>> =
         Triple("OpenLink", { OpenLinkTest() }, OpenLinkTestDoc.javaClass.name),
         Triple("Variant *", { VariantAsteriskTest() }, VariantAsteriskTestDoc.javaClass.name),
         Triple("Alignment", { AlignmentTest() }, AlignmentTestDoc.javaClass.name),
-        //        Triple("Battleship", { BattleshipTest() }, BattleshipDoc.javaClass.name),
+        Triple("Battleship", { BattleshipTest() }, BattleshipDoc.javaClass.name),
         Triple("H Constraints", { HConstraintsTest() }, ConstraintsDoc.javaClass.name),
         Triple("V Constraints", { VConstraintsTest() }, ConstraintsDoc.javaClass.name),
         Triple("Interaction", { InteractionTest() }, InteractionTestDoc.javaClass.name),
