+++ integration-tests/validation/src/testDebug/kotlin/com/android/designcompose/testapp/validation/RenderAllExamples.kt
@@ -30,8 +30,8 @@ import com.android.designcompose.DocRenderStatus
 import com.android.designcompose.TestUtils
 import com.android.designcompose.docClassSemanticsKey
 import com.android.designcompose.docRenderStatusSemanticsKey
+import com.android.designcompose.test.internal.interFont
 import com.android.designcompose.testapp.validation.examples.EXAMPLES
-import com.android.designcompose.testapp.validation.ui.interFont
 import com.github.takahirom.roborazzi.RoborazziRule
 import java.io.File
 import org.junit.BeforeClass
