+++ settings.gradle.kts
@@ -63,6 +63,9 @@ project(":tutorial-app").projectDir = File("reference-apps/tutorial/app")
 include(":battleship-app")
 
 project(":battleship-app").projectDir = File("benchmarks/battleship/app")
+
+include(":benchmarks:battleship:lib")
+
 include(":benchmarks:battleship:benchmark")
 
 include(":test:internal")
