+++ test/README
@@ -1,1 +1,1 @@
-This directory is a placeholder for a future public test support library
+This directory is a placeholder for a future public test support library.
